pint: Try to fix number reading bugs
To help track them down, I replaced a lot of internal
intepreter errors with exceptions.  This cleans up the
code quite a bit, too.

There are still some weird cases where, for instance,
inputting '10.5' when reading an integer will read 10
then send .5 to the next read.  I'm not sure at the
moment if this is a divergence from original PFC, and will
need to research.